refactor(protractor.conf): move to config
