+++ protractor.conf.js
@@ -1,48 +1,3 @@
 // @AngularClass
-require('ts-node/register');
-
-exports.config = {
-  baseUrl: 'http://localhost:3000/',
-
-  // use `npm run e2e`
-  specs: [
-    'src/**/**.e2e.ts',
-    'src/**/*.e2e.ts'
-  ],
-  exclude: [],
-
-  framework: 'jasmine2',
-
-  allScriptsTimeout: 110000,
-
-  jasmineNodeOpts: {
-    showTiming: true,
-    showColors: true,
-    isVerbose: false,
-    includeStackTrace: false,
-    defaultTimeoutInterval: 400000
-  },
-  directConnect: true,
-
-  capabilities: {
-    'browserName': 'chrome',
-    'chromeOptions': {
-      'args': ['show-fps-counter=true']
-    }
-  },
-
-  onPrepare: function() {
-    browser.ignoreSynchronization = true;
-  },
-
-  seleniumServerJar: "node_modules/protractor/selenium/selenium-server-standalone-2.48.2.jar",
-
-  /**
-   * Angular 2 configuration
-   *
-   * useAllAngular2AppRoots: tells Protractor to wait for any angular2 apps on the page instead of just the one matching
-   * `rootEl`
-   *
-   */
-   useAllAngular2AppRoots: true
-};
+// look in ./config for protractor.conf.js
+exports.config = require('./config/protractor.conf.js').config;
